refactor disco tool
