+++ scripts/disco_lib/commands/__init__.py
@@ -8,5 +8,10 @@ from .serial import serial
 from .repl import repl
 from .doctor import doctor
 from .quickstart import quickstart
+from .cables import cables
+from .check import check
 
-__all__ = ["ocd", "cpu", "mem", "flash", "serial", "repl", "doctor", "quickstart"]
+__all__ = [
+    "ocd", "cpu", "mem", "flash", "serial", "repl",
+    "doctor", "quickstart", "cables", "check",
+]
